requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
Pillow==10.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
drf-spectacular==0.27.0